+++ fews_3di/simulation.py
@@ -1,29 +1,8 @@
-"""
-Notes
-=====
-
-Importing openapi_client: that way you can only have one generated
-openapi_client. So not lizard and 3di next to each other?
-
-threedi_api_client looks like a bit of a mess. ThreediApiClient doesn't have
-an init, but a __new__. And it doesn't return a ThreediApiCLient, but
-something else. ThreediApiClient doesn't even inherit from that other
-thingy... That's not something that mypy with its proper type hints is going
-to like very much...
-
-APIConfiguration is at least a wrapper around Configuration. But somehow it
-generates an api_client inside _get_api_tokens(), which it passes part of its
-own configuration.... That looks terribly unclean.
-
-Probably too much is happening. Configuration belongs in the apps that use
-it. Some helper is OK, but it took me half an hour to figure out what was
-happening where...
-
-"""
-
 from collections import namedtuple
 from fews_3di import utils
 from pathlib import Path
+from threedi_api_client import openapi
+from threedi_api_client import ThreediApi
 from threedigrid.admin.gridresultadmin import GridH5ResultAdmin
 from typing import List
 from typing import Tuple
@@ -31,11 +10,11 @@ from typing import Tuple
 import datetime
 import logging
 import netCDF4
-import openapi_client
 import pandas as pd
 import requests
 import socket
 import time
+import warnings
 
 
 OffsetAndValue = namedtuple("OffsetAndValue", ["offset", "value"])
@@ -47,10 +26,6 @@ SIMULATION_STATUS_CHECK_INTERVAL = 30
 USER_AGENT = "fews-3di (https://github.com/nens/fews-3di/)"
 
 logger = logging.getLogger(__name__)
-
-
-class AuthenticationError(Exception):
-    pass
 
 
 class NotFoundError(Exception):
@@ -83,15 +58,12 @@ class ThreediSimulation:
     """
 
     allow_missing_saved_state: bool
-    api_client: openapi_client.ApiClient
-    configuration: openapi_client.Configuration
+    api: ThreediApi
     output_dir: Path
     saved_state_id: int
     settings: utils.Settings
     simulation_id: int
     simulation_url: str
-    simulations_api: openapi_client.SimulationsApi
-    threedimodels_api: openapi_client.ThreedimodelsApi
 
     def __init__(
         self, settings: utils.Settings, allow_missing_saved_state: bool = False
@@ -99,44 +71,18 @@ class ThreediSimulation:
         """Set up a 3di API connection."""
         self.settings = settings
         self.allow_missing_saved_state = allow_missing_saved_state
-        self.configuration = openapi_client.Configuration(host=self.settings.api_host)
-        self.api_client = openapi_client.ApiClient(self.configuration)
-        self.api_client.user_agent = USER_AGENT  # Let's be neat.
+        self.api = ThreediApi(config=self.settings.as_api_config())
+        self.api.user_agent = USER_AGENT  # Let's be neat.
         self.output_dir = self.settings.base_dir / "output"
         self.output_dir.mkdir(exist_ok=True)
         # You need to call login() and run(), but we won't: it makes testing easier.
 
     def login(self):
-        """Log in and set the necessary tokens.
-
-        Should be called right after initialising the class. It is a separate
-        method to make testing easier.
-
-        """
-        logger.info(
-            "Logging in on %s as user %s...",
-            self.settings.api_host,
-            self.settings.username,
+        warnings.warn(
+            "login() isn't needed anymore. threedi-api-client automatically "
+            "logs in when needed.",
+            DeprecationWarning,
         )
-        auth_api = openapi_client.AuthApi(self.api_client)
-        user_plus_password = openapi_client.Authenticate(
-            username=self.settings.username, password=self.settings.password
-        )
-        try:
-            tokens = auth_api.auth_token_create(user_plus_password)
-        except openapi_client.exceptions.ApiException as e:
-            status = getattr(e, "status", None)
-            if status == 401:
-                msg = (
-                    f"Authentication of '{self.settings.username}' failed on "
-                    f"{self.settings.api_host} with the given password"
-                )
-                raise AuthenticationError(msg) from e
-            logger.debug("Error isn't a 401, so we re-raise it.")
-            raise
-        # Set tokens on the configuration (which is used by self.api_client).
-        self.configuration.api_key["Authorization"] = tokens.access
-        self.configuration.api_key_prefix["Authorization"] = "Bearer"
 
     def run(self):
         """Main method
@@ -152,8 +98,6 @@ class ThreediSimulation:
         variables. Otherwise methods become harder to test in isolation.
 
         """
-        self.simulations_api = openapi_client.SimulationsApi(self.api_client)
-        self.threedimodels_api = openapi_client.ThreedimodelsApi(self.api_client)
         model_id = self._find_model()
         self.simulation_id, self.simulation_url = self._create_simulation(model_id)
 
@@ -242,7 +186,7 @@ class ThreediSimulation:
         logger.debug(
             "Searching model based on revision=%s...", self.settings.modelrevision
         )
-        threedimodels_result = self.threedimodels_api.threedimodels_list(
+        threedimodels_result = self.api.threedimodels_list(
             slug__contains=self.settings.modelrevision
         )
         results = threedimodels_result.results
@@ -265,7 +209,7 @@ class ThreediSimulation:
         data["duration"] = str(self.settings.duration)
         logger.debug("Creating simulation with these settings: %s", data)
 
-        simulation = self.simulations_api.simulations_create(data)
+        simulation = self.api.simulations_create(data)
         logger.info("Simulation %s has been created", simulation.url)
         return simulation.id, simulation.url
 
@@ -276,7 +220,7 @@ class ThreediSimulation:
 
         for name, timeserie in laterals.items():
             first_offset = timeserie[0].offset  # TODO: by definition, this is 0???
-            lateral = self.simulations_api.simulations_events_lateral_timeseries_create(
+            lateral = self.api.simulations_events_lateral_timeseries_create(
                 simulation_pk=self.simulation_id,
                 data={
                     "offset": first_offset,
@@ -293,10 +237,8 @@ class ThreediSimulation:
         while True:
             time.sleep(2)
             for id in still_to_process:
-                lateral = (
-                    self.simulations_api.simulations_events_lateral_timeseries_read(
-                        simulation_pk=self.simulation_id, id=id
-                    )
+                lateral = self.api.simulations_events_lateral_timeseries_read(
+                    simulation_pk=self.simulation_id, id=id
                 )
                 if lateral.state.lower() == "processing":
                     logger.debug("Lateral %s is still being processed.", lateral.url)
@@ -316,18 +258,23 @@ class ThreediSimulation:
                 return
 
     def _add_last_available_state(self, model_id):
-        states_result = self.threedimodels_api.threedimodels_saved_states_list(model_id)
+        states_result = self.api.threedimodels_saved_states_list(model_id)
         results = states_result.results
         if not results:
-            raise NotFoundError(f"State for model id:{model_id} not found")
+            msg = f"No saved states for model id:{model_id} found"
+            if self.allow_missing_saved_state:
+                logger.warn(msg + ", continuing")
+                return
+            else:
+                raise NotFoundError(msg)
         saved_state_id = results[0].id
         logger.info("last available state is: %s", saved_state_id)
         try:
-            self.simulations_api.simulations_initial_saved_state_create(
+            self.api.simulations_initial_saved_state_create(
                 self.simulation_id, data={"saved_state": saved_state_id}
             )
             return
-        except openapi_client.exceptions.ApiException as e:
+        except openapi.exceptions.ApiException as e:
             if e.status == 400:
                 logger.debug("Saved state setting error: %s", str(e))
                 msg = (
@@ -359,15 +306,16 @@ class ThreediSimulation:
                 state_file,
             )
             try:
-                self.simulations_api.simulations_initial_saved_state_create(
+                self.api.simulations_initial_saved_state_create(
                     self.simulation_id, data={"saved_state": saved_state_id}
                 )
                 return
-            except openapi_client.exceptions.ApiException as e:
+            except openapi.exceptions.ApiException as e:
                 if e.status == 400:
                     logger.debug("Saved state setting error: %s", str(e))
                     msg = (
-                        f"Setting initial state to saved state id={saved_state_id} failed. "
+                        f"Setting initial state to saved state "
+                        f"id={saved_state_id} failed. "
                         f"The error response was {e.body}, perhaps use "
                         f"--allow-missing-saved-state initially?"
                     )
@@ -388,7 +336,7 @@ class ThreediSimulation:
         else:
             save_time = self.settings.duration
 
-        saved_state = self.simulations_api.simulations_create_saved_states_timed_create(
+        saved_state = self.api.simulations_create_saved_states_timed_create(
             self.simulation_id,
             data={
                 "name": self.settings.simulationname,
@@ -403,29 +351,23 @@ class ThreediSimulation:
         """Upload initial waterlevel raster and wait for it to be processed."""
         logger.info("Uploading initial waterlevel raster...")
         self.waterlevel_raster_id = (
-            self.threedimodels_api.threedimodels_initial_waterlevels_list(model_id)
-            .results[0]
-            .id
+            self.api.threedimodels_initial_waterlevels_list(model_id).results[0].id
         )
 
-        ini_wl_api_call = (
-            self.simulations_api.simulations_initial2d_water_level_raster_create(
-                simulation_pk=self.simulation_id,
-                data={
-                    "aggregation_method": self.settings.initial_waterlevel,
-                    "initial_waterlevel": self.waterlevel_raster_id,
-                },
-            )
+        ini_wl_api_call = self.api.simulations_initial2d_water_level_raster_create(
+            simulation_pk=self.simulation_id,
+            data={
+                "aggregation_method": self.settings.initial_waterlevel,
+                "initial_waterlevel": self.waterlevel_raster_id,
+            },
         )
         logger.info("Added initial waterlevel raster to %s", ini_wl_api_call.url)
 
     def _add_netcdf_rain(self, rain_raster_netcdf: Path):
         """Upload rain raster netcdf file and wait for it to be processed."""
         logger.info("Uploading rain rasters...")
-        rain_api_call = (
-            self.simulations_api.simulations_events_rain_rasters_netcdf_create(
-                self.simulation_id, data={"filename": rain_raster_netcdf.name}
-            )
+        rain_api_call = self.api.simulations_events_rain_rasters_netcdf_create(
+            self.simulation_id, data={"filename": rain_raster_netcdf.name}
         )
         log_url = rain_api_call.put_url.split("?")[0]  # Strip off aws credentials.
         with rain_raster_netcdf.open("rb") as f:
@@ -436,17 +378,18 @@ class ThreediSimulation:
         logger.debug("Waiting for rain raster to be processed...")
         while True:
             time.sleep(2)
-            upload_status = (
-                self.simulations_api.simulations_events_rain_rasters_netcdf_list(
-                    self.simulation_id
-                )
+            upload_status = self.api.simulations_events_rain_rasters_netcdf_list(
+                self.simulation_id
             )
             state = upload_status.results[0].file.state
             if state.lower() == "processing":
                 logger.debug("Rain raster is still being processed.")
                 continue
             elif state.lower() == "invalid":
-                msg = f"Rain raster upload (to {log_url}) is invalid according to the server."
+                msg = (
+                    f"Rain raster upload (to {log_url}) is invalid according "
+                    f"to the server."
+                )
                 raise InvalidDataError(msg)
             elif state.lower() == "error":
                 state_description = upload_status.results[0].file.state_description
@@ -462,7 +405,7 @@ class ThreediSimulation:
         """Upload constant rainfall and wait for it to be processed."""
         logger.info("Uploading constant rainfall")
         duration = self.settings.end - self.settings.start
-        const_rain = openapi_client.models.ConstantRain(
+        const_rain = openapi.models.ConstantRain(
             simulation=self.simulation_id,
             offset=0,
             duration=int(duration.total_seconds()),
@@ -470,29 +413,14 @@ class ThreediSimulation:
             units="m/s",
         )
 
-        self.simulations_api.simulations_events_rain_constant_create(
-            self.simulation_id, const_rain
-        )
-
-    ## -------------------------------------------------------------------------------------##
-    ## function for add_design_rain for future implementation
-    # def _add_design_rain(self):
-    # """Upload design rainfall and wait for it to be processed."""
-    # logger.info("Uploading design rainfall")
-    # rain_api_call = (
-    # self.simulations_api.simulations_events_rain_rasters_lizard_create(
-    # self.simulation_id, data={
-    # 'duration': (self.settings.end - self.settings.start).total_seconds,
-    # 'values': self.settings.rain_input, #m/s , verschil tussen start en eind in secondes
-    # 'units': 'm/s'}
-    # )
+        self.api.simulations_events_rain_constant_create(self.simulation_id, const_rain)
 
     def _add_radar_rain(self):
         """Upload radar rainfall from Lizard and wait for it to be processed."""
         logger.info("Uploading radar rainfall")
         duration = self.settings.end - self.settings.start
 
-        self.simulations_api.simulations_events_rain_rasters_lizard_create(
+        self.api.simulations_events_rain_rasters_lizard_create(
             self.simulation_id,
             data={
                 "offset": 0,
@@ -507,7 +435,7 @@ class ThreediSimulation:
         """Upload rain csv timeseries and wait for them to be processed."""
         logger.info("Uploading %s rain csv timeseries...")
 
-        rain_api_call = self.simulations_api.simulations_events_rain_timeseries_create(
+        rain_api_call = self.api.simulations_events_rain_timeseries_create(
             simulation_pk=self.simulation_id,
             data={
                 "offset": rain[0],  # offset calculated in utils.py
@@ -522,7 +450,7 @@ class ThreediSimulation:
         """Upload evaporation raster netcdf file and wait for it to be processed."""
         logger.info("Uploading evaporation rasters...")
         evaporation_api_call = (
-            self.simulations_api.simulations_events_sources_sinks_rasters_netcdf_create(
+            self.api.simulations_events_sources_sinks_rasters_netcdf_create(
                 self.simulation_id, data={"filename": evaporation_raster_netcdf.name}
             )
         )
@@ -537,7 +465,7 @@ class ThreediSimulation:
         logger.debug("Waiting for evaporation raster to be processed...")
         while True:
             time.sleep(2)
-            upload_status = self.simulations_api.simulations_events_sources_sinks_rasters_netcdf_list(
+            upload_status = self.api.simulations_events_sources_sinks_rasters_netcdf_list(  # noqa: E501
                 self.simulation_id
             )
             state = upload_status.results[0].file.state
@@ -545,7 +473,10 @@ class ThreediSimulation:
                 logger.debug("Evaporation raster is still being processed.")
                 continue
             elif state.lower() == "invalid":
-                msg = f"Evaporation raster upload (to {log_url}) is invalid according to the server."
+                msg = (
+                    f"Evaporation raster upload (to {log_url}) is invalid according "
+                    f"to the server."
+                )
                 raise InvalidDataError(msg)
             elif state.lower() == "error":
                 state_description = upload_status.results[0].file.state_description
@@ -560,18 +491,14 @@ class ThreediSimulation:
     def _run_simulation(self):
         """Start simulation and wait for it to finish."""
         start_data = {"name": "queue"}
-        self.simulations_api.simulations_actions_create(
-            self.simulation_id, data=start_data
-        )
+        self.api.simulations_actions_create(self.simulation_id, data=start_data)
         logger.info("Simulation %s has been started.", self.simulation_url)
 
         start_time = time.time()
         while True:
             time.sleep(SIMULATION_STATUS_CHECK_INTERVAL)
             try:
-                simulation_status = self.simulations_api.simulations_status_list(
-                    self.simulation_id
-                )
+                simulation_status = self.api.simulations_status_list(self.simulation_id)
             except socket.gaierror as e:
                 logger.debug(e)
                 logger.warning("Hopefully temporary local network hickup")
@@ -592,7 +519,7 @@ class ThreediSimulation:
 
     def _download_results(self):
         logger.info("Downloading results into %s...", self.output_dir)
-        simulation_results = self.simulations_api.simulations_results_files_list(
+        simulation_results = self.api.simulations_results_files_list(
             self.simulation_id
         ).results
         logger.debug("All simulation results: %s", simulation_results)
@@ -608,7 +535,7 @@ class ThreediSimulation:
             if desired_result not in available_results:
                 logger.error("Desired result file %s isn't available.", desired_result)
                 continue
-            resource = self.simulations_api.simulations_results_files_download(
+            resource = self.api.simulations_results_files_download(
                 available_results[desired_result].id, self.simulation_id
             )
             target = self.output_dir / desired_result
@@ -621,7 +548,10 @@ class ThreediSimulation:
             expected_size = resource.size
             actual_size = target.stat().st_size
             if expected_size != actual_size:
-                msg = f"Incomplete download of {resource.get_url}: expected {expected_size}, got {actual_size}."
+                msg = (
+                    f"Incomplete download of {resource.get_url}: "
+                    f"expected {expected_size}, got {actual_size}."
+                )
                 raise utils.FileDownloadException(msg)
 
     def _process_basic_lizard_results(self):
@@ -633,7 +563,7 @@ class ThreediSimulation:
         if self.settings.lizard_results_scenario_uuid:
             data["result_uuid"] = self.settings.lizard_results_scenario_uuid
 
-        self.simulations_api.simulations_results_post_processing_lizard_basic_create(
+        self.api.simulations_results_post_processing_lizard_basic_create(
             simulation_pk=self.simulation_id, data=data
         )